delay fixed
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@2078 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/libsidplay/common/Mixer.java
@@ -64,14 +64,14 @@ public interface Mixer {
 	void setBalance(int sidNum, float balance);
 
 	/**
-	 * Delay feature: Delaying SID chip sound samples by time in seconds
+	 * Delay feature: Delaying SID chip sound samples by time in milliseconds
 	 * 
 	 * @param sidNum
 	 *            SID chip number
 	 * @param delay
-	 *            delay in s
+	 *            delay in ms
 	 */
-	void setDelay(int sidNum, float delay);
+	void setDelay(int sidNum, int delay);
 
 	/**
 	 * Pause player.
